refactor(store): flatten store/ into store.rs
Concatenate store/{store, raw, stream, subscription, checkpoint}.rs into
a single store.rs at the crate root. All five primitives (Store wrapper,
RawEventStore trait, EventStream + EventStreamExt, Subscription,
CheckpointStore) belong together as the "what adapters implement" surface.

Update internal callers in repository/{event_store, zero_copy, builder}.rs
that previously imported via the now-removed submodule paths
(crate::store::raw::, crate::store::store::, crate::store::stream::) to
import from crate::store::* directly.

Public paths unchanged:
- nexus_store::store::{Store, RawEventStore, EventStream, EventStreamExt,
  Subscription, CheckpointStore}
