+++ crates/nexus-store/src/repository/builder.rs
@@ -2,8 +2,7 @@ use std::marker::PhantomData;
 
 use super::event_store::EventStore;
 use super::zero_copy::ZeroCopyEventStore;
-use crate::store::raw::RawEventStore;
-use crate::store::store::Store;
+use crate::store::{RawEventStore, Store};
 
 // ═══════════════════════════════════════════════════════════════════════════
 // NeedsCodec — compile-time guard
