+++ crates/nexus-store/src/repository/event_store.rs
@@ -5,9 +5,7 @@ use super::repository::Repository;
 use crate::codec::Codec;
 use crate::envelope::pending_envelope;
 use crate::error::{AppendError, StoreError};
-use crate::store::raw::RawEventStore;
-use crate::store::store::Store;
-use crate::store::stream::EventStream;
+use crate::store::{EventStream, RawEventStore, Store};
 use crate::upcasting::{EventMorsel, Upcaster};
 use nexus::{Aggregate, AggregateRoot, DomainEvent, EventOf, Version};
 
@@ -34,7 +32,7 @@ pub(super) fn version_to_nz32(version: Version) -> Option<NonZeroU32> {
 ///
 /// # Construction
 ///
-/// Created via [`Store::repository()`](crate::store::store::Store::repository):
+/// Created via [`Store::repository()`](crate::store::Store::repository):
 ///
 /// ```ignore
 /// let store = Store::new(backend);
