--- crates/nexus-store/src/store/mod.rs
@@ -1,15 +1,0 @@
-pub(crate) mod checkpoint;
-pub(crate) mod raw;
-#[allow(
-    clippy::module_inception,
-    reason = "Store<S> is the primary public type of the store module"
-)]
-pub(crate) mod store;
-pub(crate) mod stream;
-pub(crate) mod subscription;
-
-pub use checkpoint::CheckpointStore;
-pub use raw::RawEventStore;
-pub use store::Store;
-pub use stream::{EventStream, EventStreamExt};
-pub use subscription::Subscription;
