--- crates/nexus-store/src/store/store.rs
@@ -1,46 +1,0 @@
-use std::sync::Arc;
-
-/// Shared handle to a [`RawEventStore`](super::raw::RawEventStore) backend.
-///
-/// `Store` wraps the backend in an `Arc`, making it cheap to clone and
-/// safe to share across tasks. It carries no codec, upcaster, or
-/// aggregate binding — it is just a database handle.
-///
-/// Use [`repository()`](Store::repository) to obtain a
-/// [`RepositoryBuilder`](crate::repository::RepositoryBuilder), then
-/// configure a codec and upcaster before calling `.build()`.
-///
-/// # Example
-///
-/// ```ignore
-/// let store = Store::new(FjallStore::builder("path").open()?);
-///
-/// let orders = store.repository().codec(OrderCodec).upcaster(OrderUpcaster).build();
-/// let users  = store.repository().codec(UserCodec).build();
-/// ```
-#[derive(Debug)]
-pub struct Store<S> {
-    inner: Arc<S>,
-}
-
-impl<S> Store<S> {
-    /// Wrap a raw event store backend in a shared handle.
-    pub fn new(raw: S) -> Self {
-        Self {
-            inner: Arc::new(raw),
-        }
-    }
-
-    /// Access the underlying raw store.
-    pub(crate) fn raw(&self) -> &S {
-        &self.inner
-    }
-}
-
-impl<S> Clone for Store<S> {
-    fn clone(&self) -> Self {
-        Self {
-            inner: Arc::clone(&self.inner),
-        }
-    }
-}
